requirements.txt
jupyterlab
pyyaml
black
